tests/wpt/web-platform-tests/priority-hints/link-attr.tentative.html

<!DOCTYPE html>
<title>Priority Hints - Link element</title>
<meta name="author" title="Dominic Farolino" href="mailto:domfarolino@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<link id=link1 href=resources/stylesheet.css importance=high>
<link id=link2 href=resources/stylesheet.css importance=low>
<link id=link3 href=resources/stylesheet.css importance=auto>
<link id=link4 href=resources/stylesheet.css importance=xyz>
<link id=link5 href=resources/stylesheet.css>

<script>
  test(() => {
    assert_equals(link1.importance, "high", "high importance is a valid IDL value on the link element");
    assert_equals(link2.importance, "low", "high importance is a valid IDL value on the link element");
    assert_equals(link3.importance, "auto", "auto importance is a valid IDL value on the link element");
    assert_equals(link4.importance, "auto", "invalid importance reflects as 'auto' IDL attribute on the link element");
    assert_equals(link5.importance, "auto", "missing importance reflects as 'auto' IDL attribute on the link element");
  }, "importance attribute on <link> elements should reflect valid IDL values");

  const link = document.createElement("link");
  assert_equals(link.importance, "auto");
</script>
